+++ scrypto/src/resource/mod.rs
@@ -28,7 +28,7 @@ pub use non_fungible_address::{NonFungibleAddress, ParseNonFungibleAddressError}
 pub use non_fungible_data::NonFungibleData;
 pub use non_fungible_id::{NonFungibleId, ParseNonFungibleIdError};
 pub use proof::{ParseProofError, Proof};
-pub use proof_rule::{ProofRule, ProofRuleResource, ProofRuleResourceList, SborPath, SborRelPath};
+pub use proof_rule::{ProofRule, ProofRuleResource, ProofRuleResourceList, SborPath};
 pub use resource_builder::{ResourceBuilder, DIVISIBILITY_MAXIMUM, DIVISIBILITY_NONE};
 pub use resource_def::{ParseResourceDefIdError, ResourceDef, ResourceDefId};
 pub use resource_flags::*;
